requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pydantic
mysql-connector-python
sqlalchemy
python-dotenv